BUG fixes | Please read the Description in order not to run into problems.
The ID was previously hard-coded into the app, but we changed that due to some technical difficulties we ran into. Now we hard-coded the TipUtilizator Nume. Please make sure that you have an entry in the DB with Nume = Standard and you run the sql procedure giveTipUtilizatorgetID. Cheers!
We have also reimplemented the Field Validation System on logare page.